src/requirements.txt
transformers
datasets
torch
faiss-cpu
sentence-transformers
streamlit
opencv-python
PyMuPDF
Pillow
